Move the gRPC middleware into a couchers.middleware package
Groups the interceptor stack, per-call perf accounting, API rate limiting, proto
annotation lookups and the descriptor pool into one package instead of loose
modules in the source root, and splits the sensitive-field redaction out of
interceptors.py.
