+++ app/backend/src/couchers/middleware/interceptors.py
@@ -1,19 +1,16 @@
 import logging
 from collections.abc import Callable, Mapping
-from copy import deepcopy
 from dataclasses import dataclass, field
 from datetime import datetime
-from functools import cache
 from os import getpid
 from threading import get_ident
 from time import perf_counter_ns
 from traceback import format_exception
-from typing import Any, NoReturn, cast, overload
+from typing import Any, NoReturn, cast
 from zoneinfo import ZoneInfo
 
 import grpc
 import sentry_sdk
-from google.protobuf.descriptor import Descriptor
 from google.protobuf.descriptor_pool import DescriptorPool
 from google.protobuf.message import Message
 from opentelemetry import trace
@@ -34,7 +31,6 @@ from couchers.constants import (
 )
 from couchers.context import CouchersContext, make_interactive_context, make_media_context
 from couchers.db import session_scope
-from couchers.descriptor_pool import get_descriptor_pool
 from couchers.i18n import LocalizationContext
 from couchers.metrics import (
     observe_api_call,
@@ -45,13 +41,15 @@ from couchers.metrics import (
     observe_in_servicer_setup_errors_counter,
     observe_in_servicer_setup_histogram,
 )
-from couchers.middleware_errors import CallRejectedError
+from couchers.middleware.descriptor_pool import get_descriptor_pool
+from couchers.middleware.errors import CallRejectedError
+from couchers.middleware.perf import PerfResult, read_perf, start_perf
+from couchers.middleware.proto_annotations import find_auth_level
+from couchers.middleware.ratelimit import should_rate_limit
+from couchers.middleware.sanitize import sanitized_bytes
 from couchers.models import APICall, ClientPlatform, User, UserActivity, UserSession
-from couchers.perf import PerfResult, read_perf, start_perf
 from couchers.proto import annotations_pb2
 from couchers.proto.annotations_pb2 import AuthLevel
-from couchers.proto_annotations import find_auth_level
-from couchers.ratelimit import should_rate_limit
 from couchers.utils import (
     create_lang_cookie,
     create_session_cookies,
@@ -228,80 +226,6 @@ def unauthenticated_handler[T, R](
     return abort_handler(message, status_code)
 
 
-@cache
-def _descriptor_has_sensitive(descriptor: Descriptor) -> bool:
-    """Whether this message type transitively contains any field marked sensitive."""
-    seen: set[Descriptor] = set()
-    stack = [descriptor]
-    while stack:
-        d = stack.pop()
-        if d in seen:
-            continue
-        seen.add(d)
-        for f in d.fields:
-            if f.GetOptions().Extensions[annotations_pb2.sensitive]:
-                return True
-            if f.message_type is not None:
-                stack.append(f.message_type)
-    return False
-
-
-@dataclass(frozen=True, slots=True)
-class _SanitizePlan:
-    fields_to_clear: tuple[str, ...]
-    fields_to_recurse: tuple[tuple[str, bool], ...]  # (field name, is_repeated)
-
-
-@cache
-def _sanitize_plan(descriptor: Descriptor) -> _SanitizePlan:
-    """For a message type, the fields to clear and the subfields worth recursing into."""
-    clear = []
-    recurse = []
-    for f in descriptor.fields:
-        if f.GetOptions().Extensions[annotations_pb2.sensitive]:
-            clear.append(f.name)
-        elif f.message_type is not None and _descriptor_has_sensitive(f.message_type):
-            recurse.append((f.name, f.is_repeated))
-    return _SanitizePlan(fields_to_clear=tuple(clear), fields_to_recurse=tuple(recurse))
-
-
-def _sanitize_message(message: Message) -> None:
-    plan = _sanitize_plan(message.DESCRIPTOR)
-    for name in plan.fields_to_clear:
-        message.ClearField(name)
-    for name, is_repeated in plan.fields_to_recurse:
-        submessage = getattr(message, name)
-        if not submessage:
-            continue
-        if is_repeated:
-            for msg in submessage:
-                _sanitize_message(msg)
-        else:
-            _sanitize_message(submessage)
-
-
-@overload
-def _sanitized_bytes(proto: Message) -> bytes: ...
-@overload
-def _sanitized_bytes(proto: None) -> None: ...
-def _sanitized_bytes(proto: Message | None) -> bytes | None:
-    """
-    Remove fields marked sensitive and return serialized bytes.
-
-    Sensitivity is static per message type, so the descriptor analysis is cached: messages whose type has no
-    sensitive field anywhere serialize directly without a copy or walk.
-    """
-    if not proto:
-        return None
-
-    if not _descriptor_has_sensitive(proto.DESCRIPTOR):
-        return proto.SerializeToString()
-
-    new_proto = deepcopy(proto)
-    _sanitize_message(new_proto)
-    return new_proto.SerializeToString()
-
-
 def _log_call(
     *,
     method: str,
@@ -321,8 +245,8 @@ def _log_call(
     duration = (perf_counter_ns() - start) / 1e6  # ms
     metric_method = NONEXISTENT_METHOD_LABEL if nonexistent_method else method
 
-    req_bytes = _sanitized_bytes(request)
-    res_bytes = _sanitized_bytes(response)
+    req_bytes = sanitized_bytes(request)
+    res_bytes = sanitized_bytes(response)
     response_truncated = False
     truncate_res_bytes_length = 16 * 1024  # 16 kB
     if res_bytes and len(res_bytes) > truncate_res_bytes_length:
