+++ docs/life-of-an-rpc.md
@@ -63,7 +63,7 @@ Here is a conceptual explanation of how the API system services a `GetAccountInf
 
 1. For an RPC call to hit the backend, it must be sent to the port that one of the gRPC servers is listening on.
 2. Once the server receives the request, it passes it on to each of the *interceptors* of the server, defined in `//app/backend/src/app.py`. Each interceptor receives the RPC and does something with it, either cancelling the call (e.g. if the user is not logged in) or passing it on to the next interceptor.
-3. The main interceptor that runs is the `CouchersMiddlewareInterceptor` (in `//app/backend/src/couchers/interceptors.py`) that looks up the API key in the call (from a cookie in the *metadata*, which are the HTTP headers) and makes sure it's valid, as well as looks up the user id corresponding to the key. The user id is injected into the `context` (an object passed down to the method handler) as `user_id`. The interceptor will cancel the call and return an error if no valid API key is found in the cookies.
+3. The main interceptor that runs is the `CouchersMiddlewareInterceptor` (in `//app/backend/src/couchers/middleware/interceptors.py`) that looks up the API key in the call (from a cookie in the *metadata*, which are the HTTP headers) and makes sure it's valid, as well as looks up the user id corresponding to the key. The user id is injected into the `context` (an object passed down to the method handler) as `user_id`. The interceptor will cancel the call and return an error if no valid API key is found in the cookies.
 4. Once the authenticators have each been executed and the RPC is allowed to continue, gRPC routes it to the right class method in Python. This is done using the lookup table created at startup (see the Routing section above). In this case, the full name of the RPC call is `org.couchers.api.account.Account/GetAccountInfo`, and it's routed to the `couchers.servicers.account.Account` class instance instantiated in `//app/backend/src/app.py`. In particular, the method `GetAccountInfo` is called in `//app/backend/src/couchers/servicers/account.py`.
 5. gRPC also looks up the request type and deserializes the incoming request, turning it into a Python object. In particular, gRPC calls the `empty_pb2.Empty.FromString` function on the incoming message (this doesn't do much).
 6. The `GetAccountInfo` in `account.py` (found in step 4) is actually invoked. The method signature is `def GetAccountInfo(self, request, context)`. `self` is just a reference to the class instance, and we never use it for antyhing. `request` is the `empty_pb2.Empty` Python object created (in step 5), containing the request fields. `context` contains some contextual info for the RPC, including two bits of interest: a `context.abort` function that allows aborting the RPC (erroring out with an error code and message), and the `context.user_id` with the calling user's id (from step 3).
